README.md - Remove mention of 'compute-intensive'
The mention of 'compute-intensive' brings to mind algorithms like scrypt (https://en.wikipedia.org/wiki/Scrypt) that are designed to be resistant to acceleration and used with a large number of iterations to force clients to expend computational effort. Such strategies are useful in preventing brute force attacks. See SQRL https://www.grc.com/sqrl/SQRL_Explained.pdf page 10 for example. 

This doesn't appear to be the case with casbin-server so we should clarify that Casbin-server is performing the policy enforcement check and retrieving data from elsewhere. This appears to be more in line with what is occurring and avoids pulling in extra implications of being intentionally 'compute-intensive'.